Fixed DirectFB Porter-Duff compositing example.
df_porter demonstrates the various compositing modes detailed in Thomas Porter &
Tom Duff's classic 1984 paper "Compositing Digital Images". In it, the two Toms
discuss the 12 possible binary compositing operators involving two images, A and
B. The example code does not actually demonstrate this. In the paper, sample
output is given for each operator in the four possible areas of overlap: only A,
only B, neither A nor B, and both A and B. The example code only performs
compositing in the A area, leaving areas which are only covered by B or covered
by neither uncomposited. By only rendering a rectangle with a constant alpha,
the per-pixel nature of porter-duff compositing is not demonstrated.

This change draws the A rectangle (blue) into a temporary surface, and then
composites the entire surface over B, so that the red and blue rectangles
partially overlap, so that all four areas are visible.

+-------------------------+
|                         |
|      B only             | Neither A nor B
|                         |
|               +---------+----------------+
|               |         |                |
|               | A and B |                |
|               |         |                |
+-------------------------+                |
                |                          |
Neither A nor B |        A only            |
                |                          |
                +--------------------------+

It can be more clearly seen that the XOR operator is the sum of A out B and
B out A operators.